requirements.txt
bonsai-cli
bonsai-python
